+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Wired `do_i_need_to_upgrade` into `ftplib_gui` startup and shutdown so update notices appear in the terminal
+- Added integrated `ftplib_gui upgrade` and `ftplib_gui check-updates` subcommands
+
 ## [0.1.0] - 2026-05-02
 
 ### Added